+++ App_Function_Libraries/Gradio_Related.py
@@ -42,8 +42,8 @@ from App_Function_Libraries.Gradio_UI.Plaintext_tab_import import create_plain_t
 from App_Function_Libraries.Gradio_UI.Podcast_tab import create_podcast_tab
 from App_Function_Libraries.Gradio_UI.RAG_QA_Chat_tab import create_rag_qa_chat_tab
 from App_Function_Libraries.Gradio_UI.Re_summarize_tab import create_resummary_tab
-from App_Function_Libraries.Gradio_UI.Search_Tab import create_prompt_view_tab, create_prompt_search_tab, \
-    create_search_summaries_tab, create_viewing_tab, create_search_tab, create_view_all_with_versions_tab
+from App_Function_Libraries.Gradio_UI.Search_Tab import create_prompt_search_tab, \
+    create_search_summaries_tab, create_search_tab
 from App_Function_Libraries.Gradio_UI.RAG_Chat_tab import create_rag_tab
 from App_Function_Libraries.Gradio_UI.Embeddings_tab import create_embeddings_tab, create_view_embeddings_tab, \
     create_purge_embeddings_tab
@@ -55,6 +55,7 @@ from App_Function_Libraries.Gradio_UI.Video_transcription_tab import create_vide
 from App_Function_Libraries.Gradio_UI.View_tab import create_manage_items_tab
 from App_Function_Libraries.Gradio_UI.Website_scraping_tab import create_website_scraping_tab
 from App_Function_Libraries.Gradio_UI.Chat_Workflows import chat_workflows_tab
+from App_Function_Libraries.Gradio_UI.View_DB_Items_tab import create_prompt_view_tab, create_viewing_tab, create_view_all_with_versions_tab
 #
 # Gradio UI Imports
 from App_Function_Libraries.Gradio_UI.Evaluations_Benchmarks_tab import create_geval_tab, create_infinite_bench_tab
